+++ lib/flutter_sentry.dart
@@ -53,17 +53,13 @@ class FlutterSentry {
   /// Note that this function calls for [FlutterSentry.initialize], and
   /// therefore cannot be used more than once, or in combination with
   /// [FlutterSentry.initialize].
+  ///
+  /// flutter_driver users: keep `enableFlutterDriverExtension()` call outside
+  /// of the `wrap()`.
   static T wrap<T>(
     T Function() f, {
     @required String dsn,
   }) {
-    var environment = 'debug';
-    if (kReleaseMode) {
-      environment = 'release';
-    } else if (kProfileMode) {
-      environment = 'profile';
-    }
-
     var printing = false;
     return runZoned<T>(
       () {
@@ -73,7 +69,10 @@ class FlutterSentry {
         // zone as the app: https://github.com/flutter/flutter/issues/42682.
         initialize(
           dsn: dsn,
-          environmentAttributes: Event(environment: environment),
+          environmentAttributes: const Event(
+            environment:
+                kReleaseMode ? 'release' : kProfileMode ? 'profile' : 'debug',
+          ),
         );
 
         FlutterError.onError = (details) {
@@ -149,6 +148,10 @@ class FlutterSentry {
     final event = Event(
       exception: exception,
       stackTrace: stackTrace,
+      // We should not call isFlutterDriver too early, so we don't do it inside
+      // wrap() or initialize(). Default to null, which will be substituted for
+      // what environmentAttributes provide.
+      environment: contexts_cache.isFlutterDriver() ? 'driver' : null,
       release: _sentry.environmentAttributes?.release ??
           contexts_cache.defaultReleaseString(),
       breadcrumbs: breadcrumbs.breadcrumbs.toList(),
@@ -190,6 +193,9 @@ class FlutterSentry {
   /// This event should contain static values that do not change from
   /// event to event, for example, app environment (debug, production, profile),
   /// the version of Dart/Flutter SDK, etc.
+  ///
+  /// flutter_driver users: make sure to call `enableFlutterDriverExtension()`
+  /// before `initialize()`.
   static void initialize({@required String dsn, Event environmentAttributes}) {
     if (_instance == null) {
       _instance = FlutterSentry._(